requirements.txt
Package                   Version
------------------------- --------------
discord.py                2.3.2
python-dotenv             1.0.1
openai                    0.28.0
requests                  2.31.0